✨ [feat] #6 SwiftData 마이그레이션 V2 진행
