+++ Chalkak/App/AppTypes.swift
@@ -1,0 +1,12 @@
+//
+//  AppTypes.swift
+//  Chalkak
+//
+//  Created by 배현진 on 10/4/25.
+//
+
+// 앱 전역에서 기존 네이밍 사용 (마이그레이션을 위함)
+typealias Guide         = SchemaV2.Guide
+typealias Project       = SchemaV2.Project
+typealias CameraSetting = SchemaV2.CameraSetting
+typealias Clip          = SchemaV2.Clip
