+++ Chalkak/App/ChalkakApp.swift
@@ -15,12 +15,19 @@ struct ChalkakApp: App {
     @StateObject private var coordinator = Coordinator()
     
     init() {
-        let config = ModelConfiguration()
-        self.sharedContainer = try! ModelContainer(
-            for: Clip.self, Guide.self, Project.self,
-            configurations: config
-        )
+        do {
+          self.sharedContainer = try ModelContainer(
+            for: SchemaV2.Clip.self, SchemaV2.Guide.self, SchemaV2.Project.self, SchemaV2.CameraSetting.self,
+            migrationPlan: MigrationPlan.self
+          )
+        } catch {
+          assertionFailure("ModelContainer init error: \(error)")
+          fatalError()
+        }
+        
         SwiftDataManager.shared.configure(container: sharedContainer)
+        
+        backfillGuideWasMirroredIfNeeded(container: sharedContainer)
         
         Task { @MainActor in
             SwiftDataManager.shared.cleanupAllTempProjects()
@@ -67,5 +74,36 @@ struct ChalkakApp: App {
             .environmentObject(coordinator)
         }
         .modelContainer(sharedContainer)
+    }
+    
+    private func backfillGuideWasMirroredIfNeeded(container: ModelContainer) {
+        let flagKey = "didBackfill_GuideWasMirroredAtCapture_v2_0_0"
+        guard !UserDefaults.standard.bool(forKey: flagKey) else { return }
+
+        let context = ModelContext(container)
+
+        do {
+            // 아직 false인 것만 대상(초기 기본값 false)
+            var fd = FetchDescriptor<SchemaV2.Guide>()
+            fd.predicate = #Predicate<SchemaV2.Guide> { $0.wasMirroredAtCapture == false }
+            let guides = try context.fetch(fd)
+
+            // 프로젝트 미리 로드해서 메모리 매칭
+            let projects = try context.fetch(FetchDescriptor<SchemaV2.Project>())
+
+            for guide in guides {
+                if let project = projects.first(where: { $0.guide.clipID == guide.clipID }),
+                   let cam = project.cameraSetting,
+                   cam.isFrontPosition
+                {
+                    guide.wasMirroredAtCapture = true
+                }
+            }
+
+            try context.save()
+            UserDefaults.standard.set(true, forKey: flagKey)
+        } catch {
+            print("Backfill failed: \(error)")
+        }
     }
 }
